requirements.txt
# python_version==3.11
Django~=5.0.4
djangorestframework~=3.15.1
opencv-python~=4.9.0.80
celery==5.4.0
celery[redis]
platformio~=6.1.15
pyserial==3.5
